+++ core/src/main/java/dev/iiprocraft/sg/base/adapters/SlimeAdapter.java
@@ -1,7 +1,8 @@
 /*
  * MIT License
  *
- * Copyright (c) 2021 iiProCraft & contributors
+ * Copyright (c) 2021 iiProCraft
+ * Copyright (c) 2021 Invvk
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,7 +23,7 @@
  * SOFTWARE.
  */
 
-package dev.iiprocraft.sg.api.arena.adapter.impl;
+package dev.iiprocraft.sg.base.adapters;
 
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.grinderwolf.swm.api.SlimePlugin;
